backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1

PyMySQL==1.1.0
SQLAlchemy==2.0.23

python-dotenv==1.0.1
Werkzeug==3.0.1
